[interface] change marker_size type from int to float

+++ uplot/engine/MatplotFigure.py
@@ -56,7 +56,7 @@ class MatplotFigure(IFigure):
                    color       : str | None = None,
                    line_style  : LineStyle | None = None,
                    marker_style: MarkerStyle | None = None,
-                   marker_size : int | None = None,
+                   marker_size : float | None = None,
                    opacity     : float = 1.0,
                    legend_group: str | None = None,
                    **kwargs) -> IFigure:
@@ -99,7 +99,7 @@ class MatplotFigure(IFigure):
                       name        : str | None = None,
                       color       : str | list[str] | None = None,
                       marker_style: MarkerStyle | None = None,
-                      marker_size : int | None = None,
+                      marker_size : float | None = None,
                       opacity     : float = 1.0,
                       legend_group: str | None = None,
                       **kwargs) -> IFigure:
